Fix incorrect disk usage with upper_bound function
The upper_bound function returns a list of the amount of space used
on each disk. Unfortunately, it gets this wrong: it tends to overallocate,
as it's not dealing with the rounding-down effect of the integer division
in the various bound calculations.

Modify this to use the actual amount of space allocated to the disks, and
spread that space out over the disks as evenly as possible.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>