Fix overestimation of stripe width
With five disks, it was possible to get a stripe width of 8. This is
clearly wrong, and was shown up by the "equal vs max" tests. The problem
is that function calls bind more strongly than (*), so we were multiplying
our min disks by the replication factor.

Add extra brackets to compute the max possible stripe width correctly.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>