Fix an infinite-recursion bug
We correctly pass an (decreasing) ordered list of disks to the usage
estimator, but it may return a list of disks' free space which is not
similarly ordered. For example, c=2, s=2, p=0 with five disks:

[5, 5, 5, 5, 5] -> [0, 0, 0, 0, 1]

So, we need to be able to run each iteration of the process with a sorted
disk list.

Use the process_ordered helper function to achieve this.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>