Allow deployment of the app on an arbitrary path
The app parses its URL to get the query string out, but this could be on
any arbitrary path. To use the Url.Parser code, we need to have a known
path structure to match on. So, before using the Parser, replace the
path with a known empty one. This allows us to ignore it correctly,
and makes the app deployable anywhere.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>