Correctly handle minimum stripe width larger than the number of disks
If we had no allocations at all (usually with the min stripe width being
larger than the number of disks), then the allocations array is empty,
and we don't get the right shape when it's transposed for display.

Handle the case of an empty list of allocations and return the right
shaped list of empty lists, so that the UI elements that we still need
can be rendered.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>